pyproject.toml
[project]
name = "FocusTUI"
version = "0.1.0"
authors = [
  { name="Simon Piechutowski", email="szymonpiechutowski@gmail.com" },
]
dependencies = [
    "textual~=0.73.0",
    "pygame-ce~=2.5.0",
    "platformdirs~=4.2.2",
    "python-dotenv~=1.0.1",
    "pydantic~=2.8.2",
]
description = "Your Deep Focus Session Buddy."
readme = "readme.md"
requires-python = ">=3.12"

classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
dev = [
    "textual-dev~=1.5.1",
    "ruff~=0.6.1"
]

[project.urls]
Homepage = "https://github.com/Zimzozaur/FocusTUI"
Issues = "https://github.com/Zimzozaur/FocusTUI/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
paths = ["src/focustui/static/**"]

[tool.hatch.build.targets.wheel]
paths = ["src/focustui/static/**"]

[project.scripts]
focustui = "focustui.__main__:main"

[tool.ruff]
exclude = ["src/focustui/assets.py"]

[tool.ruff.lint]
select = [
    'ALL'
]
ignore = [
    "A002",
    "ARG002",
    "ANN00",
    "ANN201",
    "D100",
    "D101",
    "D102",
    "D104",
    "D105",
    "D107",
    "D203",
    "D205",
    "D213",
    "DTZ005",
    "FBT001",
    "FBT003",
    "RUF012",
    "SIM117"
]